requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.1
blis==1.2.0
build==1.2.2.post1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
courlan==1.3.2
cryptography==44.0.0
cssselect==1.2.0
cymem==2.0.11
dataclasses-json==0.6.7
dateparser==1.2.0
debugpy==1.8.12
decorator==5.1.1
discord.py==2.4.0
distro==1.9.0
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.8
filetype==1.2.0
fpdf==1.7.2
frozenlist==1.5.0
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
htmldate==1.9.3
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jusText==3.0.1
langchain==0.3.18
langchain-anthropic==0.3.7
langchain-community==0.3.17
langchain-core==0.3.34
langchain-openai==0.3.4
langchain-pinecone==0.2.2
langchain-tests==0.3.10
langchain-text-splitters==0.3.6
langcodes==3.5.0
langdetect==1.0.9
langgraph==0.2.70
langgraph-checkpoint==2.0.12
langgraph-checkpoint-sqlite==2.0.3
langgraph-sdk==0.1.51
langsmith==0.3.4
language_data==1.3.0
lxml==5.3.0
lxml_html_clean==0.4.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.61.0
orjson==3.10.15
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pinecone==5.4.2
pinecone-client==5.0.1
pinecone-notebooks==0.0.2rc3
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pip-tools==7.4.1
platformdirs==4.3.6
pluggy==1.5.0
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pypdf==5.2.0
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.1.28
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.1
readability-lxml==0.8.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
syrupy==4.8.1
tavily-python==0.5.1
tenacity==9.0.0
thinc==8.3.4
tiktoken==0.8.0
tld==0.13
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
unstructured==0.16.17
unstructured-client==0.29.0
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
